Extract replay line trail code to a hook using useFrame
