+++ app/components/viewer/ReplayLines.tsx
@@ -1,7 +1,9 @@
-import { useFrame } from '@react-three/fiber';
-import React, { useCallback, useMemo, useRef } from 'react';
+import React, {
+    useCallback, useEffect, useMemo, useRef,
+} from 'react';
 import * as THREE from 'three';
 import { ReplayData } from '../../lib/api/apiRequests';
+import useUpdateReplayLineTrail from '../../lib/hooks/viewer/replayLines/useUpdateReplayLineTrail';
 import {
     accelerationReplayColors,
     defaultReplayColors,
@@ -11,8 +13,6 @@ import {
     inputReplayColors,
     addAlphaChannel,
 } from '../../lib/replays/replayLineColors';
-import GlobalTimeLineInfos from '../../lib/singletons/timeLineInfos';
-import { getSampleIndexNearTime, getSampleNearTime } from '../../lib/utils/replay';
 import ReplayChartHoverLocation from './ReplayChartHoverLocation';
 import ReplayDnf from './ReplayDnf';
 import ReplayGears from './ReplayGears';
@@ -34,12 +34,6 @@ export const LineTypes: { [name: string]: LineType } = {
     inputs: { name: 'Inputs', colorsCallback: inputReplayColors },
 };
 
-interface LineUpdateIndices {
-    startIndex: number;
-    endIndex: number;
-    segmentUpdate: boolean;
-}
-
 interface ReplayLineProps {
     replay: ReplayData;
     lineType: LineType;
@@ -55,10 +49,6 @@ const ReplayLine = ({
         return addAlphaChannel(colors, 0);
     }, [replay, lineType]);
 
-    const previousLineUpdate = useRef<LineUpdateIndices>();
-
-    const timeLineGlobal = GlobalTimeLineInfos.getInstance();
-
     const onUpdate = useCallback(
         (self) => {
             self.setFromPoints(points);
@@ -67,86 +57,14 @@ const ReplayLine = ({
         [points, colorBuffer],
     );
 
-    useFrame(() => {
+    // Update the replay line trail based on trail settings
+    const { manualLineTrailUpdate } = useUpdateReplayLineTrail(bufferGeom, replay, TRAIL_FADE_SEGMENT_TIME);
+
+    useEffect(() => {
         if (!bufferGeom.current) return;
-
-        const {
-            showFullTrail, showTrailToStart, revealTrailTime, currentRaceTime,
-        } = timeLineGlobal;
-
-        const curSampleIndex = getSampleIndexNearTime(replay, timeLineGlobal.currentRaceTime);
-
-        if (showFullTrail || showTrailToStart) {
-            if (!previousLineUpdate.current || previousLineUpdate.current.segmentUpdate) {
-                // Reset alpha of full line if it's the first update or if the previous update was a segment update
-                for (let i = 0; i < points.length; i++) {
-                    bufferGeom.current.attributes.color.setW(i, 1);
-                }
-                bufferGeom.current.attributes.color.needsUpdate = true;
-            }
-
-            if (showFullTrail) {
-                bufferGeom.current.setDrawRange(0, points.length);
-            } else {
-                bufferGeom.current.setDrawRange(0, curSampleIndex);
-            }
-
-            previousLineUpdate.current = {
-                startIndex: 0,
-                endIndex: points.length,
-                segmentUpdate: false,
-            };
-        } else {
-            // Get trail indices
-            const startTrailIndex = getSampleIndexNearTime(
-                replay,
-                currentRaceTime - revealTrailTime,
-            );
-            const endFadeIndex = getSampleIndexNearTime(
-                replay,
-                currentRaceTime - revealTrailTime + TRAIL_FADE_SEGMENT_TIME,
-            );
-
-            // Clamp trail indices
-            const startTrailIndexClamped = Math.max(0, Math.min(startTrailIndex, replay.samples.length));
-            const endFadeIndexClamped = Math.max(0, Math.min(endFadeIndex, curSampleIndex, replay.samples.length));
-
-            const prev = previousLineUpdate.current;
-            // Update starts at trail index or previous start index (or 0 if no previous update has occured)
-            const updateStart = Math.min(startTrailIndexClamped, prev?.startIndex || 0);
-            // Update starts at sample index or previous end index (or end of line if no previous update has occured)
-            const updateEnd = Math.max(curSampleIndex, prev?.endIndex || points.length);
-
-            // Update line alpha if needed (previous was not a segment update or update range changed)
-            if (!prev?.segmentUpdate || updateStart !== prev?.startIndex || updateEnd !== prev?.endIndex) {
-                for (let i = updateStart; i < updateEnd; i++) {
-                    if (i < startTrailIndexClamped) {
-                        // Index before minimum index, hide line: set alpha to 0
-                        bufferGeom.current.attributes.color.setW(i, 0);
-                    } else if (i < endFadeIndexClamped) {
-                        // Index after trail start, before fade end, fade alpha between 0 and 1
-                        const alpha = (i - startTrailIndexClamped) / (endFadeIndexClamped - startTrailIndexClamped);
-                        bufferGeom.current.attributes.color.setW(i, alpha);
-                    } else {
-                        // Index after fade end, show line: set alpha to 1
-                        bufferGeom.current.attributes.color.setW(i, 1);
-                    }
-                }
-                bufferGeom.current.attributes.color.needsUpdate = true;
-            }
-
-            // Set line draw range
-            const samplesToDraw = curSampleIndex - startTrailIndexClamped;
-            bufferGeom.current.setDrawRange(startTrailIndexClamped, samplesToDraw);
-
-            // Set start and end to trail start and end
-            previousLineUpdate.current = {
-                startIndex: startTrailIndexClamped,
-                endIndex: curSampleIndex,
-                segmentUpdate: true,
-            };
-        }
-    });
+        bufferGeom.current.setAttribute('color', colorBuffer);
+        manualLineTrailUpdate();
+    }, [colorBuffer, manualLineTrailUpdate]);
 
     return (
         <line>
